+++ CHANGELOG.md
@@ -122,7 +122,7 @@
 
 - add bias metrics (holisticai.metrics.bias)
 - add mitigation techniques (holisticai.mitigation.bias)
-- add bias plotting tools (holisticai.bias.plots)
+- add bias plotting tools (holisticai.plots.bias)
 - add tutorials for measuring bias (tutorials/measuring_bias_tutorials)
 - add tutorials for mitigating bias (tutorials/mitigating_bias_tutorials)